Use immutableSet utility to set style properties
