+++ packages/block-editor/src/components/global-styles/color-panel.js
@@ -28,6 +28,7 @@ import { __ } from '@wordpress/i18n';
 import ColorGradientControl from '../colors-gradients/control';
 import { useColorsPerOrigin, useGradientsPerOrigin } from './hooks';
 import { getValueFromVariable } from './utils';
+import { immutableSet } from '../../utils/object';
 
 export function useHasColorPanel( settings ) {
 	const hasTextPanel = useHasTextPanel( settings );
@@ -328,10 +329,13 @@ export default function ColorPanel( {
 	const userTextColor = decodeValue( value?.color?.text );
 	const hasTextColor = () => !! userTextColor;
 	const setTextColor = ( newColor ) => {
-		onChange( {
-			...value,
-			color: { ...value?.color, text: encodeColorValue( newColor ) },
-		} );
+		onChange(
+			immutableSet(
+				value,
+				[ 'color', 'text' ],
+				encodeColorValue( newColor )
+			)
+		);
 	};
 	const resetTextColor = () => setTextColor( undefined );
 
@@ -343,34 +347,31 @@ export default function ColorPanel( {
 	const userGradient = decodeValue( value?.color?.gradient );
 	const hasBackground = () => !! userBackgroundColor || !! userGradient;
 	const setBackgroundColor = ( newColor ) => {
-		onChange( {
-			...value,
-			color: {
-				...value?.color,
-				background: encodeColorValue( newColor ),
-				gradient: undefined,
-			},
-		} );
+		const newValue = immutableSet(
+			value,
+			[ 'color', 'background' ],
+			encodeColorValue( newColor )
+		);
+		newValue.color.gradient = undefined;
+		onChange( newValue );
 	};
 	const setGradient = ( newGradient ) => {
-		onChange( {
-			...value,
-			color: {
-				...value?.color,
-				background: undefined,
-				gradient: encodeGradientValue( newGradient ),
-			},
-		} );
+		const newValue = immutableSet(
+			value,
+			[ 'color', 'gradient' ],
+			encodeGradientValue( newGradient )
+		);
+		newValue.color.background = undefined;
+		onChange( newValue );
 	};
 	const resetBackground = () => {
-		onChange( {
-			...value,
-			color: {
-				...value?.color,
-				background: undefined,
-				gradient: undefined,
-			},
-		} );
+		const newValue = immutableSet(
+			value,
+			[ 'color', 'background' ],
+			undefined
+		);
+		newValue.color.gradient = undefined;
+		onChange( newValue );
 	};
 
 	// Links
@@ -380,19 +381,13 @@ export default function ColorPanel( {
 	);
 	const userLinkColor = decodeValue( value?.elements?.link?.color?.text );
 	const setLinkColor = ( newColor ) => {
-		onChange( {
-			...value,
-			elements: {
-				...value?.elements,
-				link: {
-					...value?.elements?.link,
-					color: {
-						...value?.elements?.link?.color,
-						text: encodeColorValue( newColor ),
-					},
-				},
-			},
-		} );
+		onChange(
+			immutableSet(
+				value,
+				[ 'elements', 'link', 'color', 'text' ],
+				encodeColorValue( newColor )
+			)
+		);
 	};
 	const hoverLinkColor = decodeValue(
 		inheritedValue?.elements?.link?.[ ':hover' ]?.color?.text
@@ -401,46 +396,28 @@ export default function ColorPanel( {
 		value?.elements?.link?.[ ':hover' ]?.color?.text
 	);
 	const setHoverLinkColor = ( newColor ) => {
-		onChange( {
-			...value,
-			elements: {
-				...value?.color?.elements,
-				link: {
-					...value?.elements?.link,
-					':hover': {
-						...value?.elements?.link?.[ ':hover' ],
-						color: {
-							...value?.elements?.link?.[ ':hover' ]?.color,
-							text: encodeColorValue( newColor ),
-						},
-					},
-				},
-			},
-		} );
+		onChange(
+			immutableSet(
+				value,
+				[ 'elements', 'link', ':hover', 'color', 'text' ],
+				encodeColorValue( newColor )
+			)
+		);
 	};
 	const hasLink = () => !! userLinkColor || !! userHoverLinkColor;
-	const resetLink = () =>
-		onChange( {
-			...value,
-			elements: {
-				...value?.color?.elements,
-				link: {
-					...value?.color?.elements?.link,
-					color: {
-						...value?.color?.elements?.link?.color,
-						text: undefined,
-					},
-					':hover': {
-						...value?.color?.elements?.link?.[ ':hover' ],
-						color: {
-							...value?.color?.elements?.link?.[ ':hover' ]
-								?.color,
-							text: undefined,
-						},
-					},
-				},
-			},
-		} );
+	const resetLink = () => {
+		let newValue = immutableSet(
+			value,
+			[ 'elements', 'link', ':hover', 'color', 'text' ],
+			undefined
+		);
+		newValue = immutableSet(
+			newValue,
+			[ 'elements', 'link', 'color', 'text' ],
+			undefined
+		);
+		onChange( newValue );
+	};
 
 	// Elements
 	const elements = [
@@ -614,69 +591,42 @@ export default function ColorPanel( {
 				elementGradientUserColor
 			);
 		const resetElement = () => {
-			onChange( {
-				...value,
-				elements: {
-					...value?.elements,
-					[ name ]: {
-						...value?.elements?.[ name ],
-						color: {
-							...value?.elements?.[ name ]?.color,
-							background: undefined,
-							gradient: undefined,
-							text: undefined,
-						},
-					},
-				},
-			} );
+			const newValue = immutableSet(
+				value,
+				[ 'elements', name, 'color', 'background' ],
+				undefined
+			);
+			newValue.elements[ name ].color.gradient = undefined;
+			newValue.elements[ name ].color.text = undefined;
+			onChange( newValue );
 		};
 
-		const setElementTextColor = ( newValue ) => {
-			onChange( {
-				...value,
-				elements: {
-					...value?.elements,
-					[ name ]: {
-						...value?.elements?.[ name ],
-						color: {
-							...value?.elements?.[ name ]?.color,
-							text: encodeColorValue( newValue ),
-						},
-					},
-				},
-			} );
+		const setElementTextColor = ( newTextColor ) => {
+			onChange(
+				immutableSet(
+					value,
+					[ 'elements', name, 'color', 'text' ],
+					encodeColorValue( newTextColor )
+				)
+			);
 		};
-		const setElementBackgroundColor = ( newValue ) => {
-			onChange( {
-				...value,
-				elements: {
-					...value?.elements,
-					[ name ]: {
-						...value?.elements?.[ name ],
-						color: {
-							...value?.elements?.[ name ]?.color,
-							background: encodeColorValue( newValue ),
-							gradient: undefined,
-						},
-					},
-				},
-			} );
+		const setElementBackgroundColor = ( newBackgroundColor ) => {
+			const newValue = immutableSet(
+				value,
+				[ 'elements', name, 'color', 'background' ],
+				encodeColorValue( newBackgroundColor )
+			);
+			newValue.elements[ name ].color.gradient = undefined;
+			onChange( newValue );
 		};
-		const setElementGradient = ( newValue ) => {
-			onChange( {
-				...value,
-				elements: {
-					...value?.elements,
-					[ name ]: {
-						...value?.elements?.[ name ],
-						color: {
-							...value?.elements?.[ name ]?.color,
-							gradient: encodeGradientValue( newValue ),
-							background: undefined,
-						},
-					},
-				},
-			} );
+		const setElementGradient = ( newGradient ) => {
+			const newValue = immutableSet(
+				value,
+				[ 'elements', name, 'color', 'gradient' ],
+				encodeGradientValue( newGradient )
+			);
+			newValue.elements[ name ].color.background = undefined;
+			onChange( newValue );
 		};
 		const supportsTextColor = true;
 		// Background color is not supported for `caption`
