+++ packages/block-editor/src/hooks/test/utils.js
@@ -7,112 +7,8 @@ import { applyFilters } from '@wordpress/hooks';
  * Internal dependencies
  */
 import '../anchor';
-import { immutableSet } from '../utils';
 
 const noop = () => {};
-
-describe( 'immutableSet', () => {
-	describe( 'handling falsy values properly', () => {
-		it( 'should create a new object if `undefined` is passed', () => {
-			const result = immutableSet( undefined, 'test', 1 );
-
-			expect( result ).toEqual( { test: 1 } );
-		} );
-
-		it( 'should create a new object if `null` is passed', () => {
-			const result = immutableSet( null, 'test', 1 );
-
-			expect( result ).toEqual( { test: 1 } );
-		} );
-
-		it( 'should create a new object if `false` is passed', () => {
-			const result = immutableSet( false, 'test', 1 );
-
-			expect( result ).toEqual( { test: 1 } );
-		} );
-
-		it( 'should create a new object if `0` is passed', () => {
-			const result = immutableSet( 0, 'test', 1 );
-
-			expect( result ).toEqual( { test: 1 } );
-		} );
-
-		it( 'should create a new object if an empty string is passed', () => {
-			const result = immutableSet( '', 'test', 1 );
-
-			expect( result ).toEqual( { test: 1 } );
-		} );
-
-		it( 'should create a new object if a NaN is passed', () => {
-			const result = immutableSet( NaN, 'test', 1 );
-
-			expect( result ).toEqual( { test: 1 } );
-		} );
-	} );
-
-	describe( 'manages data assignment properly', () => {
-		it( 'assigns value properly when it does not exist', () => {
-			const result = immutableSet( {}, 'test', 1 );
-
-			expect( result ).toEqual( { test: 1 } );
-		} );
-
-		it( 'overrides existing values', () => {
-			const result = immutableSet( { test: 1 }, 'test', 2 );
-
-			expect( result ).toEqual( { test: 2 } );
-		} );
-
-		describe( 'with array notation access', () => {
-			it( 'assigns values at deeper levels', () => {
-				const result = immutableSet( {}, [ 'foo', 'bar', 'baz' ], 5 );
-
-				expect( result ).toEqual( { foo: { bar: { baz: 5 } } } );
-			} );
-
-			it( 'overrides existing values at deeper levels', () => {
-				const result = immutableSet(
-					{ foo: { bar: { baz: 1 } } },
-					[ 'foo', 'bar', 'baz' ],
-					5
-				);
-
-				expect( result ).toEqual( { foo: { bar: { baz: 5 } } } );
-			} );
-
-			it( 'keeps other properties intact', () => {
-				const result = immutableSet(
-					{ foo: { bar: { baz: 1 } } },
-					[ 'foo', 'bar', 'test' ],
-					5
-				);
-
-				expect( result ).toEqual( {
-					foo: { bar: { baz: 1, test: 5 } },
-				} );
-			} );
-		} );
-	} );
-
-	describe( 'does not mutate the original object', () => {
-		it( 'clones the object at the first level', () => {
-			const input = {};
-			const result = immutableSet( input, 'test', 1 );
-
-			expect( result ).not.toBe( input );
-		} );
-
-		it( 'clones the object at deeper levels', () => {
-			const input = { foo: { bar: { baz: 1 } } };
-			const result = immutableSet( input, [ 'foo', 'bar', 'baz' ], 2 );
-
-			expect( result ).not.toBe( input );
-			expect( result.foo ).not.toBe( input.foo );
-			expect( result.foo.bar ).not.toBe( input.foo.bar );
-			expect( result.foo.bar.baz ).not.toBe( input.foo.bar.baz );
-		} );
-	} );
-} );
 
 describe( 'anchor', () => {
 	const blockSettings = {
